Cleared out some bitrot.  To be honest there looks like a lot of changes but they are ancient.  I don't think I have looked that this in over two years.  It does compile and run now, though.
Changed the function call in LAMBDA-IN-DYNAMIC-ENVIRONMENT from SPLIT-IF to
SPLIT-ON, which is due to a change I made in TOOLBOX recently.  Still not using
the TOOLBOX implementation of that, don't know why.

Fixed the name on the thread so that it is a string instead of a symbol.
Bordeaux-Threads apparently doesn't find the symbol name anymore.

WHILE is no longer exported from the TOOLBOX library and is as deprecated as
things get.  Should switch to ITERATE.